Fix for wrong dates
Hiho,
when i replace the news repository all dates lose one day. I tested this on two different systems. This is my fix for it. It seems that the timestamp in the db is saved as UTC, so i lose 1 h in germany and a date like 14.1.16 12 am become 13.1.16 23pm. Then you combine the date timestamp with the time timestamp and cast it as datetime without adding the right timezone. Usually this is made by the fluid viewhelper format.date but if he gets a datetime he doesn't add the timezone and in our case its useless anyway because after the combine we will falsify the event time.

I hope i could help you out.
Thx for the great extension.