[TASK] Restrict scheduled CI run to main repo
In the GitHub CI file, a scheduled run is configured, so that
the checks and tests run once a day. This is a good idea, to
change possible problems due to updated depencies early.

However, the scheduled run also runs in the forks and this
is unnecessary. The checks will run in all forks of contributors
(in case they have enabled this in general) and they will get
notifcations in case of failures. This is unnecessary, it is
enough for the scheduled runs to run in one repo.